Scenes, menus, and folders
- Created base level scene to load different levels onto additively.
- Added death screen, pending retry and main menu functionality.
- Organized scripts into folders.
- Prevent projectiles from dealing damage to original shooter.
- Other various changes.

+++ Assets/Scripts/UI/HUD.cs
@@ -60,7 +60,7 @@ public class HUD : MonoBehaviour
     private void LateUpdate()
     {
         // health info
-        healthNumber.text = playerHealth.currentHealth.ToString() + "/" + playerHealth.maxHealth.ToString();
+        healthNumber.text = ((int)playerHealth.currentHealth).ToString() + "/" + playerHealth.maxHealth.ToString();
         healthBar.SetSizeWithCurrentAnchors(RectTransform.Axis.Horizontal, healthBarWidth * playerHealth.currentHealth / playerHealth.maxHealth);
 
         // healing info
